requirements.txt
django
gunicorn
django-heroku
psycopg2
djangorestframework
django-cors-headers
datetimerange
whitenoise
requests
